+++ src/browser-ext/js/utils.js
@@ -10,35 +10,41 @@
   window.zadark = window.zadark || {}
 
   window.zadark.utils = {
-    setThemeModeAttribute: (themeMode) => {
-      document.documentElement.setAttribute('data-theme-mode', themeMode)
+    setThemeAttr: (themeMode) => {
+      document.documentElement.setAttribute('data-zadark-theme', themeMode)
     },
 
-    setPageTheme: function ({ themeMode, customTheme }) {
-      if (['custom', 'single'].includes(themeMode)) {
-        this.setThemeModeAttribute(customTheme)
-      }
+    setPageTheme: function (theme) {
+      switch (theme) {
+        case 'light':
+        case 'dark': {
+          this.setThemeAttr(theme)
+          return
+        }
 
-      if (themeMode === 'auto') {
-        const isDark = window.matchMedia('(prefers-color-scheme: dark)').matches
-        this.setThemeModeAttribute(isDark ? 'dark' : 'light')
+        case 'auto': {
+          const isDark = window.matchMedia('(prefers-color-scheme: dark)').matches
+          this.setThemeAttr(isDark ? 'dark' : 'light')
+          return
+        }
+
+        default: {
+          this.setThemeAttr('dark')
+        }
       }
     },
 
     refreshPageTheme: function () {
-      window.zadark.browser.getExtensionSettings().then(({ themeMode, customTheme }) => {
-        this.setPageTheme({
-          themeMode,
-          customTheme
-        })
+      window.zadark.browser.getExtensionSettings().then(({ theme }) => {
+        this.setPageTheme(theme)
       })
     }
   }
 
   window.matchMedia('(prefers-color-scheme: dark)').addListener(function (event) {
-    window.zadark.browser.getExtensionSettings().then(({ themeMode }) => {
-      if (themeMode === 'auto') {
-        window.zadark.utils.setThemeModeAttribute(event.matches ? 'dark' : 'light')
+    window.zadark.browser.getExtensionSettings().then(({ theme }) => {
+      if (theme === 'auto') {
+        window.zadark.utils.setThemeAttr(event.matches ? 'dark' : 'light')
       }
     })
   })